realizando carga do dataset contratos com sucesso para BD Relacional MySQL (posteriormente fazer carga em BD nao relacional) - restringindo a 100k registros, ajuste em chave identificador para BIGINT (20)

+++ src/main/resources/sql/script_contrato.sql
@@ -1,5 +1,5 @@
 CREATE TABLE contrato (  
-  identificador INT(11) NOT NULL,
+  identificador BIGINT(20) NOT NULL,
   uasg VARCHAR(30),
   codigo_contrato VARCHAR(30),
   modalidade_licitacao VARCHAR(30),
